Moved send and receive to their own directory. Removed cleanuploop stuff from send, it should be implemented around send in MIST, but not inside send. This keeps both send and receive generic (issue #6)
